requirements.txt
alembic==1.6.5
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
